upd: remove CscriptCheck from main
